refactor(fe): integrate grade page into assignment page (#2725)
* refactor(fe): integrate grade page into assignment page

* chore(fe): remove unused imports

* chore(fe): add a comment explaining component deactivation
