+++ apps/frontend/app/admin/_components/ManagementSidebar.tsx
@@ -53,17 +53,13 @@ const getCourseNavItems = (courseId: string): NavItem[] => [
     name: 'Assignment',
     path: `/admin/course/${courseId}/assignment`,
     icon: AssignmentIcon
-  },
+  }
   // {
   //   name: 'Exam',
   //   path: `/admin/course/${courseId}/exam`,
   //   icon: FaFilePen
   // },
-  {
-    name: 'Grade',
-    path: `/admin/course/${courseId}/grade`,
-    icon: GradeIcon
-  }
+
   // {
   //   name: 'Q&A',
   //   path: `/admin/course/${courseId}/qna`,
@@ -337,32 +333,6 @@ function HomeIcon(props: { className: string }) {
       xmlns="http://www.w3.org/2000/svg"
     >
       <path d="M15.0045 7.98783C15.0045 8.42587 14.6395 8.769 14.2258 8.769H13.447L13.4641 12.6675C13.4641 12.7332 13.4592 12.7989 13.4519 12.8646V13.2564C13.4519 13.7943 13.0163 14.2299 12.4785 14.2299H12.0891C12.0624 14.2299 12.0356 14.2299 12.0088 14.2274C11.9747 14.2299 11.9407 14.2299 11.9066 14.2299H11.1157H10.5317C9.99385 14.2299 9.55824 13.7943 9.55824 13.2564V12.6724V11.1149C9.55824 10.6842 9.21025 10.3362 8.77951 10.3362H7.22205C6.79131 10.3362 6.44331 10.6842 6.44331 11.1149V12.6724V13.2564C6.44331 13.7943 6.00771 14.2299 5.4699 14.2299H4.88585H4.10955C4.07305 14.2299 4.03654 14.2274 4.00004 14.225C3.97084 14.2274 3.94164 14.2299 3.91243 14.2299H3.52307C2.98526 14.2299 2.54965 13.7943 2.54965 13.2564V10.5309C2.54965 10.509 2.54965 10.4846 2.55209 10.4627V8.769H1.77092C1.33288 8.769 0.992188 8.4283 0.992188 7.98783C0.992188 7.76881 1.06519 7.57413 1.23554 7.40378L7.47513 1.96482C7.64548 1.79448 7.84017 1.77014 8.01051 1.77014C8.18086 1.77014 8.37554 1.81881 8.52156 1.94049L14.7368 7.40378C14.9315 7.57413 15.0288 7.76881 15.0045 7.98783Z" />
-    </svg>
-  )
-}
-
-function GradeIcon(props: { className: string }) {
-  return (
-    <svg
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clip-path="url(#clip0_9941_13175)">
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M9.10897 1.09429C8.49034 0.495832 7.50856 0.495832 6.88994 1.09429L6.58764 1.38673C6.23388 1.72896 5.74199 1.88878 5.25464 1.81985L4.83819 1.76095C3.98594 1.6404 3.19167 2.21748 3.04295 3.06526L2.97028 3.47954C2.88524 3.96434 2.58124 4.38276 2.14645 4.61345L1.7749 4.81058C1.01457 5.214 0.711181 6.14773 1.08918 6.92101L1.2739 7.29888C1.49006 7.74108 1.49006 8.25828 1.2739 8.70048L1.08918 9.07835C0.71118 9.85163 1.01457 10.7854 1.7749 11.1888L2.14645 11.3859C2.58124 11.6166 2.88524 12.035 2.97028 12.5198L3.04295 12.9341C3.19167 13.7819 3.98594 14.359 4.83819 14.2384L5.25464 14.1795C5.74199 14.1106 6.23389 14.2704 6.58764 14.6126L6.88994 14.9051C7.50857 15.5035 8.49034 15.5035 9.10897 14.9051L9.41127 14.6126C9.76502 14.2704 10.2569 14.1106 10.7443 14.1795L11.1607 14.2384C12.013 14.359 12.8072 13.7819 12.956 12.9341L13.0286 12.5198C13.1137 12.035 13.4177 11.6166 13.8525 11.3859L14.224 11.1888C14.9843 10.7854 15.2877 9.85163 14.9097 9.07835L14.725 8.70048C14.5089 8.25828 14.5089 7.74107 14.725 7.29888L14.9097 6.92101C15.2877 6.14772 14.9843 5.214 14.224 4.81058L13.8525 4.61345C13.4177 4.38276 13.1137 3.96433 13.0286 3.47954L12.956 3.06526C12.8072 2.21748 12.013 1.6404 11.1607 1.76095L10.7443 1.81985C10.2569 1.88878 9.76502 1.72896 9.41127 1.38673L9.10897 1.09429ZM4.80594 11.3317H6.29688L6.80306 9.79471H9.21433L9.7113 11.3317H11.193L8.89221 4.66847H7.11597L4.80594 11.3317ZM8.8646 8.71792H7.15279L7.98108 6.16861H8.0271L8.8646 8.71792Z"
-          className={props.className}
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_9941_13175">
-          <rect width="16" height="16" fill="white" />
-        </clipPath>
-      </defs>
     </svg>
   )
 }
