+++ apps/frontend/app/admin/_components/code-editor/EditorLayout.tsx
@@ -48,14 +48,16 @@ export function EditorLayout({
 
   return (
     // Admin Layout의 Sidebar를 무시하기 위한 fixed
-    <div className="grid-rows-editor fixed left-0 grid h-dvh w-full min-w-[1000px] overflow-x-auto bg-slate-800 text-white">
+    <div className="grid-rows-editor fixed left-0 top-0 grid h-dvh w-full min-w-[1000px] overflow-x-auto bg-slate-800 text-white">
       <header className="flex h-12 justify-between bg-slate-900 px-6">
         <div className="flex items-center justify-center gap-4 text-lg text-[#787E80]">
           <Link href="/">
             <Image src={codedangLogo} alt="코드당" width={33} />
           </Link>
           <Link
-            href={`/admin/course/${courseId}/grade/assignment/${assignmentId}`}
+            href={
+              `/admin/course/${courseId}/assignment/${assignmentId}` as const
+            }
           >
             {assignment.title}
           </Link>
