+++ apps/frontend/app/admin/course/[courseId]/assignment/[assignmentId]/(overall)/@tabs/assessment/user/[userId]/problem/[problemId]/layout.tsx
