--- apps/frontend/app/admin/course/[courseId]/grade/error.tsx
@@ -1,23 +1,0 @@
-'use client'
-
-import { ErrorDetail } from '@/components/ErrorDetail'
-import { captureError } from '@/libs/captureError'
-import { useEffect } from 'react'
-
-interface Props {
-  error: Error & { digest?: string }
-  reset: () => void
-}
-
-export default function Error({ error }: Props) {
-  useEffect(() => {
-    captureError(error)
-  }, [error])
-
-  return (
-    <ErrorDetail
-      errorDetail="Failed to Load Grade Detail Page!"
-      error={error}
-    />
-  )
-}
